leadhunter/backend/requirements.txt
# LeadHunter backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
apscheduler>=3.10.4
anthropic>=0.18.0
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-multipart>=0.0.6
email-validator>=2.1.0
